cleanup: break cyclic dependencies
Break the `gcs/project.py` file in two, and move `gcs/iam.py` to
`testbench/servers/`.  This breaks the cycle between the `testbench/`
and `gcs/` modules. I am still planning to reorganize the code to make
it harder to create such cycles, and to make the structure easier to
grok (or at least to explain without blushing).
