+++ testbench/rest_server.py
@@ -25,6 +25,7 @@ from werkzeug.middleware.dispatcher import DispatcherMiddleware
 from google.storage.v2 import storage_pb2
 import gcs as gcs_type
 import testbench
+from testbench.servers import iam_rest_server, projects_rest_server
 
 
 db = testbench.database.Database.init()
@@ -867,10 +868,10 @@ def delete_resumable_upload(bucket_name):
 # === SERVER === #
 
 # Define the WSGI application to handle HMAC key and service account requests
-(PROJECTS_HANDLER_PATH, projects_app) = gcs_type.project.get_projects_app(db)
+(PROJECTS_HANDLER_PATH, projects_app) = projects_rest_server.get_projects_app(db)
 
 # Define the WSGI application to handle IAM requests
-(IAM_HANDLER_PATH, iam_app) = gcs_type.iam.get_iam_app()
+(IAM_HANDLER_PATH, iam_app) = iam_rest_server.get_iam_app()
 
 server = flask.Flask(__name__)
 server.debug = False
